Refactor Header Provider Info & Details
