+++ src/components/Header/component/Layout.stories.js
@@ -3,10 +3,10 @@ import { storiesOf } from '@storybook/react'
 import * as React from 'react'
 import styles from '~/components/layout/PageFrame/index.scss'
 import Layout from './Layout'
-import ProviderInfo from './ProviderInfo'
-import ProviderDetails from './ProviderInfo/UserDetails'
-import ProviderDisconnected from './ProviderDisconnected'
-import ConnectDetails from './ProviderDisconnected/ConnectDetails'
+import ProviderAccesible from './ProviderInfo/ProviderAccesible'
+import UserDetails from './ProviderDetails/UserDetails'
+import ProviderDisconnected from './ProviderInfo/ProviderDisconnected'
+import ConnectDetails from './ProviderDetails/ConnectDetails'
 
 const FrameDecorator = story => (
   <div className={styles.frame}>
@@ -20,8 +20,8 @@ storiesOf('Components /Header', module)
     const provider = 'Metamask'
     const userAddress = '0x873faa4cddd5b157e8e5a57e7a5479afc5d30moe'
     const network = 'RINKEBY'
-    const info = <ProviderInfo provider={provider} network={network} userAddress={userAddress} connected />
-    const details = <ProviderDetails provider={provider} network={network} userAddress={userAddress} connected />
+    const info = <ProviderAccesible provider={provider} network={network} userAddress={userAddress} connected />
+    const details = <UserDetails provider={provider} network={network} userAddress={userAddress} connected />
 
     return <Layout providerInfo={info} providerDetails={details} />
   })
@@ -35,8 +35,8 @@ storiesOf('Components /Header', module)
     const provider = 'Metamask'
     const userAddress = '0x873faa4cddd5b157e8e5a57e7a5479afc5d30moe'
     const network = 'RINKEBY'
-    const info = <ProviderInfo provider={provider} network={network} userAddress={userAddress} connected={false} />
-    const details = (<ProviderDetails
+    const info = <ProviderAccesible provider={provider} network={network} userAddress={userAddress} connected={false} />
+    const details = (<UserDetails
       provider={provider}
       network={network}
       userAddress={userAddress}
